requirements.txt
Django==4.2.13
Pillow==11.0.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
qrcode==8.0